+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* API: limit and offset parameters to countries, countrycodes, languages, tags, states endpoints
 ## [0.7.12] 2021-04-22
 ### Added
 * API: multiple station endpoints do not support limit,offset and hidebroken parameters